+++ js/main.js
@@ -1,3 +1,4 @@
 //Точка входа. Модуль, который связывает другие модули
 import './popup.js';
 import './form.js';
+import './map.js';
